Get data frames of types of GxE from a mash object

Performs set operations to determine pairwise GxE for effects from a mash object.

get_GxE(m, factor = 0.4, thresh = 0.05)

Arguments

m
An object of type mash

factor
a number between 0 and 1. The factor within which effects are considered to be shared.

thresh
Numeric. The threshold for including an effect in the assessment

Value

A list containing eight data frames. Those with names that start "S_" contain significant effects of different types between pairs of named rows and columns. S_all_pairwise contains all significant effects; NS_pairwise contains all non-significant effects. S_CN contains effects significant in only one condition, and effects with a significantly different magnitude (differential sensitivity). This dataframe is not conservative using the local false sign rate test - we can't determine the sign of one of the effects for effects significant in only one condition - so it's not recommended to use this, but included. S_2_no contains effects significant in both conditions that do not differ significantly in magnitude. These effects do not have GxE. S_AP contains effects significant in both conditions that differ in their sign - and have antagonistic pleiotropy. S_DS contains effects significant in both conditions that differ in the magnitude of their effect, but not their sign - differentially sensitive alleles. S_1_row and S_1_col contain effects that are significant in just one of the two conditions - the row or the column, respectively.